+++ packages/ZbayLite/src/renderer/components/widgets/channels/ChannelsListItem.tsx
@@ -72,7 +72,7 @@ export const ChannelsListItem: React.FC<IChannelsListItemComponentProps> = ({
 }) => {
   const classes = useStyles({})
   const isFromZbay = channel.username !== 'Unknown'
-  const highlight = selected.id === channel.key
+  const highlight = selected.id === channel.key || selected.id === channel.username
   return (
     <ListItem
       button
@@ -80,7 +80,7 @@ export const ChannelsListItem: React.FC<IChannelsListItemComponentProps> = ({
       onClick={() => {
         history.push(
           `/main/${directMessages
-            ? `direct-messages/${channel.key}/${channel.username}`
+            ? `direct-messages/${channel.username}`
             : `channel/${channel.key}`
           }`
         )