+++ packages/ZbayLite/src/renderer/sagas/socket/socket.saga.test.ts
@@ -312,7 +312,7 @@ describe('checkCertificatesSaga', () => {
     const randomId = 'randomId'
     const address = 'address'
     const message = '' as IMessage
-    const encryptedMessage = 'encryptedMessage'
+    const encryptedMessage = 'keKhopVVNJtPAPBCooSZ3uB4nh8lsQ63lmOqZlRX+zUfEi2IJ/6WL5NPD+tTFw71F4ArpZJ7SgeqqPyJo6lnz2Ls1tMxZH6T/KuwJq0x70fV7M1LMqazim/LdMvqrKq0vOlwMkZbmijtH6wOalxeaA=='
     const encryptedPhrase = 'encryptedPhrase'
     const conversationId = 'conversationId'
     const contactPublicKey = 'contactPublicKey'
@@ -320,7 +320,7 @@ describe('checkCertificatesSaga', () => {
     const identityStore = new Identity()
 
     const conversation = {
-      sharedSecret: 'sharedSecret',
+      sharedSecret: '9bf9fb2db723642d6126d3fdb5df6fe374cef4c1fa1c5403dcbe849d3d30a3fd',
       contactPublicKey: contactPublicKey,
       conversationId: conversationId
     }
@@ -403,14 +403,14 @@ describe('checkCertificatesSaga', () => {
         ],
         [
           matchers.call.fn(computeSecrets),
-          'sharedSecret'
+          '9bf9fb2db723642d6126d3fdb5df6fe374cef4c1fa1c5403dcbe849d3d30a3fd'
         ],
         [
           matchers.call.fn(getPublicKey),
           conversationId
         ],
         [
-          matchers.call(encryptMessage, 'sharedSecret', 'no panicpubKey'),
+          matchers.call(encryptMessage, '9bf9fb2db723642d6126d3fdb5df6fe374cef4c1fa1c5403dcbe849d3d30a3fd', 'no panicpubKey'),
           encryptedMessage
         ],
         [
@@ -418,7 +418,7 @@ describe('checkCertificatesSaga', () => {
           'parsedMessage'
         ],
         [
-          matchers.call(encryptMessage, 'sharedSecret', 'parsedMessage'),
+          matchers.call(encryptMessage, '9bf9fb2db723642d6126d3fdb5df6fe374cef4c1fa1c5403dcbe849d3d30a3fd', 'parsedMessage'),
           encryptedPhrase
         ],
         [