+++ packages/ZbayLite/src/renderer/store/certificates/certificates.saga.ts
@@ -1,14 +1,15 @@
 import { call, apply, all, takeEvery, put, select } from 'typed-redux-saga'
 import { PayloadAction } from '@reduxjs/toolkit'
-
 import { certificatesActions } from './certificates.reducer'
 import { createUserCsr, configCrypto, CertFieldsTypes, parseCertificate, getCertFieldValue } from '@zbayapp/identity'
 import electronStore from '../../../shared/electronStore'
+import { actions } from '../handlers/directMessages'
 import { actions as identityActions } from '../handlers/identity'
 import { registrationServiceAddress } from '../../../shared/static'
-import notificationsHandlers from '../../store/handlers/notifications'
+import notificationsHandlers from '../handlers/notifications'
 import { successNotification } from '../handlers/utils'
 import directMessagesSelectors from '../selectors/directMessages'
+import contactsSelectors from '../selectors/contacts'
 
 const filterCertificates = (certificates: string[]): string[] => {
   return certificates.filter((cert) => {
@@ -29,13 +30,27 @@ export function* responseGetCertificates(
   action: PayloadAction<ReturnType<typeof certificatesActions.responseGetCertificates>['payload']>
 ): Generator {
   const certificates = action.payload
-
   const filteredCertificates = filterCertificates(certificates.certificates)
   yield* put(certificatesActions.setUsersCertificates(filteredCertificates))
+  const users = yield* select(contactsSelectors.usersCertificateMapping)
+  for (const [key, value] of Object.entries(users)) {
+    if (value.dmPublicKey && value.username) {
+      yield put(
+        actions.fetchUsers({
+          usersList: {
+            [value.username]: {
+              publicKey: key,
+              halfKey: value.dmPublicKey,
+              nickname: value.username
+            }
+          }
+        })
+      )
+    }
+  }
 }
 
 export function* responseGetCertificate(): Generator {
-  console.log('Response get cert saga, setting registration status')
   electronStore.set('isNewUser', false)
   yield* put(
     identityActions.setRegistraionStatus({
@@ -66,11 +81,9 @@ export function* createOwnCertificate(
     }
   }
 
-  const hiddenServices: HiddenServicesType = yield* apply(
-    electronStore,
-    electronStore.get,
-    ['hiddenServices']
-  )
+  const hiddenServices: HiddenServicesType = yield* apply(electronStore, electronStore.get, [
+    'hiddenServices'
+  ])
 
   const peerIdAddress = yield* apply(electronStore, electronStore.get, ['peerId'])
 
@@ -78,7 +91,6 @@ export function* createOwnCertificate(
     console.log('invalid peer id address or not exist')
     return
   }
-
   const dmPublicKey = yield* select(directMessagesSelectors.publicKey)
 
   const userData = {
@@ -90,8 +102,6 @@ export function* createOwnCertificate(
     hashAlg: configCrypto.hashAlg
   }
 
-  console.log('userData', userData)
-
   const user = yield* call(createUserCsr, userData)
 
   yield put(
@@ -102,7 +112,6 @@ export function* createOwnCertificate(
   )
 
   yield* put(certificatesActions.setOwnCertKey(user.userKey))
-  console.log('After registering csr')
 }
 
 export function* certificatesSaga(): Generator {