+++ packages/ZbayLite/src/renderer/store/handlers/directMessages.ts
@@ -40,13 +40,7 @@ export class DirectMessages {
 export type DirectMessagesStore = DirectMessages
 
 export const initialState: DirectMessagesStore = {
-  users: {
-    '02dc8264c555d46b3f6b16f1e751e979ebc69e6df6a02e7d4074a5df981e507da2': {
-      nickname: 'holmes',
-      publicKey: '02dc8264c555d46b3f6b16f1e751e979ebc69e6df6a02e7d4074a5df981e507da2',
-      halfKey: '0bfb475810c0e26c9fab590d47c3d60ec533bb3c451596acc3cd4f21602e9ad9'
-    }
-  },
+  users: {},
   conversations: {},
   conversationsList: {},
   privateKey: '',