+++ packages/ZbayLite/src/renderer/store/selectors/channel.ts
@@ -181,20 +181,20 @@ export const inputLocked = createSelector(
   waggleSelectors.isConnected,
   isPublicChannel,
   (channelId, waggleContacts, waggle, publicChannel) => {
-    const contactsData: Array<{ publicKey: string }> = Object.values(waggleContacts)
+    const contactsData: Array<{ publicKey: string; nickname: string }> = Object.values(waggleContacts)
     const currentContactArray = contactsData.filter(item => {
-      return item.publicKey === channelId
+      return item.publicKey === channelId || item.nickname === channelId
     })
 
     if (!waggle) {
       return INPUT_STATE.NOT_CONNECTED
     }
 
-    if (publicChannel || currentContactArray[0]) {
+    if (publicChannel || !currentContactArray[0]?.nickname?.startsWith('anon')) {
       return INPUT_STATE.AVAILABLE
     }
 
-    if (!currentContactArray[0]) {
+    if (!currentContactArray[0] || currentContactArray[0]?.nickname?.startsWith('anon')) {
       return INPUT_STATE.USER_NOT_REGISTERED
     }